Removes the default port (80) from the host header
Some S3 server realizations do not expect the 80 port to be
included in the host header (notable example: s3.yandexcloud.net).
For this reason Minio client was not able to authenticate on such
servers.

This change removes the port from the host header if the value
is 80. This makes the headers logic same as popular python boto
and Java S3 SDKs.